jenkins/jobs: Fix creation of script location on nodes
Nodes need to have a way to get the scripts that execute jobs updated on them.
To do this, a new builder - node-git-prep - has been added that builds up
the script location on the nodes in /srv/git. If the location doesn't exist,
it will be created and a new git checkout of the infrastructure made at
that location.

Note that this will access the repo from git.asterisk.org, to reduce contention
on the gerrit server.