jenkins/scripts/gerrit-git-prep: Pull the repo from Gerrit
OpenStack takes the approach of having the repos be cloned from another
Git repository other than Gerrit. This makes sense when you are of the
size of OpenStack; for us, it is additional complexity that is not quite
necessary. (And if it ever does become necessary, we can worry about
that it then.)

This patch updates the Git URL to point to Gerrit. Once the repo is
cloned, it then updates the repo with the changes stored in zuul-merger.

For some extra debugging on the Jenkins console, it also spits out the
most important environment variables passed to the script.