jenkins/scripts/prep-testsuite.sh: Add a script to pull the TestSuite
This patch adds a basic script that clones the Asterisk Test Suite, such
that it can be used for runs against Asterisk. Unlike the more generic
gerrit-git-prep script, this script always pulls directly from Gerrit
and doesn't bother trying to sync up with a Zuul merger repo. This is
because the intent is to just run the latest Test Suite against
Asterisk, and not to verify the Test Suite itself.