jenkins: Run the testsuite with elevated permissions
Some tests require a pcap to sniff network traffic for validation that
certain messages were sent/received. That sniffing requires elevated
permissions; hence, for all tests to actually pass, we have to run the
testsuite with elevated permissions.

This script is nice enough to set permissions to the CI user when it is
done.