build-asterisk: Execute "alembic branches" before build.
To help reduce the amount of times we have created Alembic branches
this adds branch checking to the build script. It will run before
Asterisk is even built to ensure quickly if the commit alone will
cause an Alembic branch to occur.

As of the time of this commit this is not yet run on commit, so
if multiple reviews are up an Alembic branch may still go into
the tree. Since there are nightly runs, however, those will catch
the Alembic branch overnight and it can be fixed quickly.

ASTERISK-25685 #close

Change-Id: Ib35e66c515634dc01a7de60f675100b27d0836c5