jenkins/scripts/prep-testsuite.sh: Run as root to handle off-nominals.
When a job is aborted or abnormally terminates the permissions of
some directories may be left as root instead of the jenkins user. This
results in them sticking around and sometimes becoming part of the
artifacts.

This change executes prep-testsuite.sh as root to ensure it can clean
up and also explicitly cleans out the logs directories.

Change-Id: I146853afad9c34d4dee6b887d4681b1f98de712a