jenkins/scripts: Be more invasive when checking git clones.
There are certain cases where git clones have been interrupted or ended
up corrupted. This causes subsequents jobs to fail as they do a shallow
check for the presence of a git clone before doing a pull. This change
adds a deeper check which will delete the clone if it is not in a
pristine state. This means that if a corrupted git clone does occur it
will fix itself the next time the job is used on a node.

Change-Id: Ic1d6f394073d313198d9dd7cc0e96b4cc33bfe32