Remove separate codec abi test and make build-asterisk.sh quieter
The codec abi job has failed since we renamed the codec downloads
directories.  Instead of a separate job, the unit test jobs for
non-master, non-11 branches now enable codec_silk and let the
downloader install it as asterisk is installed.  If the module loads
successfully, we're good.  To reduce the need for downloads,
--with-sounds-cache=/srv/cache/sounds and
--with-externals-cache=/srv/cache/externals are now passed to
./configure.  The directories are created if they don't exist.
This way all jobs can share the same cache.

Also, build-asterisk.sh is so chatty it makes it hard to troubleshoot.
The original behavior was that a parallel make was run first and if
it failed a serialized make was run to insure that the failure was
last in the log. Both makes used NOISY_BUILD.  To reduce the log
clutter, the parallel make is now always run without NOISY_BUILD.  If it
fails, the serialized make is run with NOISY_BUILD so the messages and
command lines appear.

Change-Id: I108bf8412b8484f4ad9be5cb30d7754b71e1958c