Limit gate-asterisk-ari to run tests in the tests/rest_api directory
gate-asterisk-ari uses a tag to limit tests rather than including only
the rest_api directory. This causes almost 900 tests to be considered
for execution when only about 160 are actually needed.   Now the
test runs only the tests in the rest_api directory.

There are only 2 tests outside that directory that have the ARI tag and
they're only tagged that way because they use ARI for the test. They're
not specifically testing ARI functionality.

Change-Id: I8dc95e9e7af094497cd95b07aed6486c93a36565