jobs:  Update node labels
Nodes are now marked with 1 or more of the following labels to indicate
the job types they are capable of running:
check debug doc gate periodic realtime

In addition to the 32/64 bit label, the nodes are also now marked with
more specific distribution and version information:
centos centos6 centos7 debian debian8 fedora fedora25 ubuntu
ubuntu14 ubuntu16

A 64 bit CentOS6 node would be marked: centos centos6 64-bit

Each of the jobs' node parameters was updated to reflect the type
of job that it is.  None of the jobs were distribution-specific so
the original "centos || debian" node labels were removed.

Change-Id: I296fd9c81474e9817e576108445657cf6e6cb17b