asterisk-jobs:  Limit MWI and ARI tests to only those applicable
The ARI and MWI tests were using the '-g' option to runtests to select
the appropriate tests to run.  Unfortunately, there are 2 side effects
to that.  The first is that runtests has to scan every testsuite test
looking for the handful to run and second is that runtests reports the
ones not run as "skipped".  This causes Jenkins to display all the
skipped tests in the results.

There were no ARI tests outside rest_api so that '-g ARI' was simply
changed to '-t tests/rest_ari'.  The MWI tests are all in a very few
directories so  '-t' options were added to limit the search to them.

Change-Id: Ia95637e1303576a7dddf3a47964dcf31deac53da